perf(ledger): snapshot view owns its data (no Arc::clone) (#702)
Root cause of the 460 ms-per-block ceiling at epoch 25+ on preview
from-genesis catch-up.

`LedgerStateSnapshot::from(&LedgerState)` was `Arc::clone`-ing six hot
Arc-shared HashMaps from the live state:
  - delegations
  - pool_params
  - reward_accounts
  - stake_key_deposits
  - governance
  - epoch_blocks_by_pool

The snapshot view is moved into a `spawn_blocking` task that holds it
for the entire bincode-serialize + fs::write cycle — typically
0.5–1 s for a 40 MB preview snapshot, longer for mainnet.  During
that window the source Arc's refcount is 2, so every subsequent
`apply_delta_to_state` call in the apply hot path that touches any of
those maps via `Arc::make_mut` triggers a CoW **deep clone of the
entire HashMap** (~10–30 MB per map at preview's epoch-29 size,
~100 MB at full mainnet stake-credential set).

Across the six maps that's the ~460 ms-per-block ceiling observed
across PIDs 10445, 68508, 84765 every time the catch-up reached
~block 119k (around epoch 25 on preview, when delegation/governance
state grows enough that the per-block CoW dominates over apply
itself, which stayed at <100 µs throughout).

## Fix

Change the snapshot view's six Arc fields to plain owned HashMaps.
The `From<&LedgerState> for LedgerStateSnapshot` impl deep-clones the
Arc contents via `(*arc).clone()` at view-build time (one-time hit
under the apply write lock — ~60 MB transient at preview scale, paid
once per snapshot instead of once per intervening block).  The
reverse direction wraps the owned maps back into `Arc::new(...)` so
the rehydrated `LedgerState` has the same CoW semantics the apply
path expects.

After this change the live `LedgerState`'s Arc refcount stays at 1
across snapshot windows, so `Arc::make_mut` is the in-place mutation
path (O(1)) rather than CoW deep-clone.

## Reference

Mirrors Haskell's `Ouroboros.Consensus.Storage.LedgerDB.V2.InMemory`
which holds independent ledger states per "anchor point" — the
LedgerDB volatile sequence and snapshot writes never share mutable
state, by construction.  Dugite's Arc-shared approach was a Rust-side
optimisation that backfired when held across spawn_blocking
boundaries.

## Tests

All 2354 dugite-ledger + dugite-node tests pass.  No semantic change
to the on-disk snapshot format (bincode encoding is identical —
HashMap and Arc<HashMap> serialise the same).

Refs #702, #698

+++ crates/dugite-node/src/node/connection_lifecycle.rs
@@ -1276,7 +1276,7 @@ impl ConnectionLifecycleManager {
                 let phase_offset = {
                     let mut h = std::collections::hash_map::DefaultHasher::new();
                     addr.hash(&mut h);
-                    std::time::Duration::from_millis((h.finish() % 200) as u64)
+                    std::time::Duration::from_millis(h.finish() % 200)
                 };
                 let mut poll_ticker = tokio::time::interval_at(
                     tokio::time::Instant::now() + phase_offset,
